Corrected error models relationships

+++ paralympics/routes.py
@@ -1,4 +1,6 @@
-from flask import current_app as app, request, make_response
+from flask import current_app as app, request, abort, jsonify
+from marshmallow.exceptions import ValidationError
+from sqlalchemy import exc
 from paralympics import db
 from paralympics.models import Region, Event
 from paralympics.schemas import RegionSchema, EventSchema
@@ -10,11 +12,40 @@ events_schema = EventSchema(many=True)
 event_schema = EventSchema()
 
 
+# Error handlers
+@app.errorhandler(ValidationError)
+def register_validation_error(error):
+    """ Error handler for marshmallow schema validation errors.
+
+    Args:
+        error (ValidationError): Marshmallow error.
+
+    Returns:
+        HTTP response with the validation error message and the 400 status code
+    """
+    response = error.messages
+    return response, 400
+
+
+@app.errorhandler(404)
+def resource_not_found(e):
+    """ Error handler for 404.
+
+        Args:
+            HTTP 404 error
+
+        Returns:
+            JSON response with the validation error message and the 404 status code
+        """
+    return jsonify(error=str(e)), 404
+
+
 @app.get("/regions")
 def get_regions():
     """Returns a list of NOC region codes and their details in JSON.
 
-    :returns: JSON
+    Returns:
+        JSON for all the regions
     """
     # Select all the regions using Flask-SQLAlchemy
     all_regions = db.session.execute(db.select(Region)).scalars()
@@ -28,23 +59,33 @@ def get_regions():
 def get_region(code):
     """ Returns one region in JSON.
 
-    :param code: The NOC code of the region to return
-    :param type code: str
-    :returns: JSON
+    Returns 404 if the region code is not found in the database.
+
+    Args:
+        code (str): The 3 digit NOC code of the region to be searched for
+
+    Returns: 
+        JSON for the region if found otherwise 404
     """
     # Query structure shown at https://flask-sqlalchemy.palletsprojects.com/en/3.1.x/queries/#select
-    region = user = db.session.execute(db.select(Region).filter_by(NOC=code)).scalar_one()
-    # Dump the data using the Marshmallow region schema; '.dump()' returns JSON.
-    result = region_schema.dump(region)
-    # Return the data in the HTTP response
-    return result
+    # Try to find the region, if it is ot found, catch the error and return 404
+    try:
+        region = db.session.execute(db.select(Region).filter_by(NOC=code)).scalar_one()
+        # Dump the data using the Marshmallow region schema; '.dump()' returns JSON.
+        result = region_schema.dump(region)
+        # Return the data in the HTTP response
+        return result
+    except exc.NoResultFound as e:
+        # See https://flask.palletsprojects.com/en/2.3.x/errorhandling/#returning-api-errors-as-json
+        abort(404, description="Region not found")
 
 
 @app.get("/events")
 def get_events():
     """Returns a list of events and their details in JSON.
 
-    :returns: JSON
+    Returns: 
+        JSON for all events
     """
     all_events = db.session.execute(db.select(Event)).scalars()
     result = events_schema.dump(all_events)
@@ -55,9 +96,11 @@ def get_events():
 def get_event(event_id):
     """ Returns the event with the given id JSON.
 
-    :param event_id: The id of the event to return
-    :param type event_id: int
-    :returns: JSON"""
+    Args:
+        event_id (int): The id of the event to return
+    Returns:
+        JSON
+    """
     event = db.session.execute(db.select(Event).filter_by(id=event_id)).scalar_one()
     result = event_schema.dump(event)
     return result
@@ -70,7 +113,8 @@ def add_event():
    Gets the JSON data from the request body and uses this to deserialise JSON to an object using Marshmallow
    event_schema.loads()
 
-   :returns: JSON
+   Returns: 
+        JSON
    """
     ev_json = request.get_json()
     event = event_schema.load(ev_json)
@@ -86,7 +130,9 @@ def add_region():
     Gets the JSON data from the request body and uses this to deserialise JSON to an object using Marshmallow
    region_schema.loads()
 
-    :returns: JSON"""
+    Returns: 
+        JSON
+    """
     json_data = request.get_json()
     region = region_schema.load(json_data)
     db.session.add(region)
@@ -98,30 +144,41 @@ def add_region():
 def delete_event(event_id):
     """ Deletes the event with the given id.
 
-    :param event_id: The id of the event to delete
-    :returns: JSON"""
+    Args: 
+        event_id (int): The id of the event to delete
+    Returns: 
+        JSON
+    """
     event = db.session.execute(db.select(Event).filter_by(id=event_id)).scalar_one()
     db.session.delete(event)
     db.session.commit()
-    return {"message": f"Event {event_id} deleted"}
+    return {"message": f"Event {event_id} deleted."}
 
 
 @app.delete('/regions/<noc_code>')
 def delete_region(noc_code):
     """ Deletes the region with the given code.
 
-    :param code: The NOC code of the region to delete
-    :returns: JSON"""
+    Args:
+        param code (str): The 3-character NOC code of the region to delete
+    Returns:
+        JSON
+    """
     region = db.session.execute(db.select(Region).filter_by(NOC=noc_code)).scalar_one()
-    db.session.delete(region)
-    db.session.commit()
-    return {"message": f"Region {noc_code} deleted"}
+    if region:
+        db.session.delete(region)
+        db.session.commit()
+        return {"message": f"Region {noc_code} deleted."}
+    else:
+        abort(404, description="Region not found")
 
 
 @app.patch("/events/<event_id>")
 def event_update(event_id):
-    """Updates changed fields for the event.
-
+    """ Updates changed fields for the specified event.
+    
+    Returns:
+        JSON message
     """
     # Find the event in the database
     existing_event = db.session.execute(
@@ -134,20 +191,20 @@ def event_update(event_id):
     # Commit the changes to the database
     db.session.add(event_update)
     db.session.commit()
-    # Return json showing the updated record
-    updated_event = db.session.execute(
-        db.select(Event).filter_by(event_id=event_id)
-    ).scalar_one_or_none()
-    result = event_schema.jsonify(updated_event)
-    response = make_response(result, 200)
-    response.headers["Content-Type"] = "application/json"
+    # Return json success message
+    response = {"message": f"Event with id={event_id} updated."}
     return response
 
 
 @app.patch("/regions/<noc_code>")
 def region_update(noc_code):
-    """Updates changed fields for the region.
+    """Updates changed fields for the specified region.
+    
+    Args:
+        noc_code (str): 3 character NOC region code
 
+    Returns:
+        JSON message
     """
     # Find the region in the database
     existing_region = db.session.execute(
@@ -160,11 +217,6 @@ def region_update(noc_code):
     # Commit the changes to the database
     db.session.add(region_update)
     db.session.commit()
-    # Return json showing the updated record
-    updated_region = db.session.execute(
-        db.select(Region).filter_by(NOC=noc_code)
-    ).scalar_one_or_none()
-    result = region_schema.jsonify(updated_region)
-    response = make_response(result, 200)
-    response.headers["Content-Type"] = "application/json"
+    # Return json message
+    response = {"message": f"Region {noc_code} updated."}
     return response
